新增校验

+++ src/main/java/com/kingstar/controller/ClbumController.java
@@ -6,6 +6,7 @@ import com.kingstar.bean.Student;
 import com.kingstar.response.Response;
 import com.kingstar.service.ClbumService;
 import io.swagger.annotations.ApiOperation;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,7 +23,10 @@ public class ClbumController {
     public Response<Page> getClbumList(@RequestBody Page page) {
         int totalRecords = clbumService.getClbumCount();
         Page curPage = new Page(page.getCurrentPageNum(), totalRecords);
-        curPage.setPageSize(page.getPageSize());
+        //校验分页大小，限定50
+        int size = Math.min(page.getPageSize(), 50);
+        if (size <= 0) size = 0;
+        curPage.setPageSize(size);
         List<Student> list = clbumService.getClbumList(curPage);
         curPage.setRecords(list);
         return new Response<>("200", "查询成功", curPage);
@@ -38,6 +42,10 @@ public class ClbumController {
     @ApiOperation("新增班级信息")
     @PostMapping("/addClbum")
     public Response<Integer> addClbum(@RequestBody Clbum clbum) {
+        if (StringUtils.isBlank(clbum.getName())
+                || clbum.getNumber() != null) {
+            throw new RuntimeException("字段非空");
+        }
         Integer result = clbumService.addClbum(clbum);
         if (result != null && result > 0) {
             return new Response<>("200", "新增班级成功", result);
@@ -48,7 +56,15 @@ public class ClbumController {
 
     @ApiOperation("修改班级信息")
     @PostMapping("/updateClbum")
-    public Response<Integer> updateClbum(@RequestBody Clbum clbum){
+    public Response<Integer> updateClbum(@RequestBody Clbum clbum) {
+        if (StringUtils.isBlank(clbum.getName())
+                || clbum.getNumber() != null) {
+            throw new RuntimeException("字段非空");
+        }
+        Clbum clbumTmp = clbumService.getClbumById(clbum.getId());
+        if (clbumTmp == null) {
+            return new Response<>("200", "不存在该班级", null);
+        }
         Integer result = clbumService.updateClbum(clbum);
         if (result != null && result > 0) {
             return new Response<>("200", "修改班级信息成功", result);
@@ -59,7 +75,11 @@ public class ClbumController {
 
     @ApiOperation("删除班级信息")
     @PostMapping("/removeClbum")
-    public Response<Integer> removeClbum(Integer id){
+    public Response<Integer> removeClbum(Integer id) {
+        Clbum clbumTmp = clbumService.getClbumById(id);
+        if (clbumTmp == null) {
+            return new Response<>("200", "不存在该班级", null);
+        }
         Integer result = clbumService.removeClbum(id);
         if (result != null && result > 0) {
             return new Response<>("200", "删除班级信息成功", result);
@@ -70,7 +90,11 @@ public class ClbumController {
 
     @ApiOperation("删除班级信息")
     @DeleteMapping("/deletedClbum")
-    public Response<Integer> deletedClbum(Integer id){
+    public Response<Integer> deletedClbum(Integer id) {
+        Clbum clbumTmp = clbumService.getClbumById(id);
+        if (clbumTmp == null) {
+            return new Response<>("200", "不存在该班级", null);
+        }
         Integer result = clbumService.deletedClbum(id);
         if (result != null && result > 0) {
             return new Response<>("200", "删除班级信息成功", result);
